+++ hsrb_gazebo_bringup/CHANGELOG.rst
@@ -2,6 +2,32 @@
 Changelog for package hsrb_gazebo_bringup
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.3.0 (2025-07-29)
+-------------------
+* simulator support for apply_force
+* add description for hsrc case
+* Changes due to ign being renamed to gz
+* Change for Ignition Gazebo
+* Add support for grasp action in Gazebo
+* Fix lint errors
+* Add hsrb_gz_ros2_control
+* Enable velocity control for the base_roll_joint
+* Move hsrb_rviz_simulator to hsrb_launch
+* add create static tf
+* update rviz
+* Modify gripper fake's position limits and parameters of apply_force action to work with the RViz simulator.
+* Added add_mesh
+* Fixed set_distance command
+* Add gazebo_plugins to hsrb_gazebo_bringup as exec_depend
+* Publish calibrated camera frames as static tfs
+* fix typo, remove unused packages
+* modify indent
+* rename controller, update gripper action
+* hsrb_gripper_bridge -> forward_command_controller
+* use hardware_interface instead of control_interface
+* update gripper interface, add hsrb_gripper_bridge
+* Contributors: Hiroaki Yaguchi, Katsushi Fukuoka, Keisuke Takeshita, Shigeo Tsuduki, Tomohiro Ono
+
 2.1.0 (2024-10-16)
 -------------------
 * Migration to Humble