+++ hsrb_gripper_fake_interface/CHANGELOG.rst
@@ -2,6 +2,14 @@
 Changelog for package hsrb_gripper_fake_interface
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.3.0 (2025-07-29)
+-------------------
+* simulator support for apply_force
+* add description for hsrc case
+* Changes due to ign being renamed to gz
+* Change for Ignition Gazebo
+* Contributors: Katsushi Fukuoka, Shigeo Tsuduki
+
 2.2.0 (2025-05-08)
 -------------------
 * Migration to ROS2 humble